Refactored example handling by consolidating file retrieval and lifecycle hook invocation
Also, to ensure the examples being validated and the one considered for duplication is the same list
